chore: run browser-less API specs once under a dedicated api project
API-tier specs assert over HTTP and never open a browser, yet every one ran
twice under the desktop and mobile projects. A third Playwright project now
runs *.api.test.ts files once and the browser projects ignore them: four
fully-API specs take the suffix, and the mixed import and shared-records
specs split into API and UI halves, with the shared payload builders moved
to e2e/import.helpers.ts. 470 executions per run become 418.
