+++ e2e/securities.api.test.ts
